Made some minor adjustment
